initial implementation of Model_LoadModel()
NOTE:
compiles
no runtime checks have been performed!
there's TODO's in the source stating what needs to be done
FS_FOpenFile and FS_FLoadFile return a size_t instead of an int now;
I believe that this commit revises the caller code
partitions driver code from refresh-definition code

comitting this since I don't know when I will be able to work on it again

+++ include/common/filesystem.h
@@ -2,9 +2,10 @@
 #define GUARD_QUAKE_COMMON_FILESYSTEM_H
 
 #include <stdio.h>
+#include <stddef.h>
 
-int FS_FOpenFile(char const *filename, FILE **file);
-int FS_FLoadFile(char const *filename, void **buffer);
+size_t FS_FOpenFile(char const *filename, FILE **file);
+size_t FS_FLoadFile(char const *filename, void **buffer);
 void FS_FCloseFile(FILE **file);
 void FS_Init(void);
